Problem: Purpose of `self._check_cassette` not immediately clear
Solution: Rename it and move to a utility module, along with the `scrub_host_name` function. Renamed `tests.py` to `test_tas_api.py` while I'm at it.
